refactor(agent): 将 Message 重构为枚举并支持工具调用
- 将 Message 结构体改为带标签的枚举，以更好地表示不同角色的消息
- 添加对工具调用（tool_calls）的支持，包括在 Assistant 消息中
- 将 ProviderError::ApiError 的 code 字段从 i32 改为 u16
- 更新 LLMNode 以正确处理 Assistant 消息的内容提取
- 修改 ChatModel 以处理新的消息枚举格式

+++ nodes/src/agent/providers/openrouter.rs
@@ -119,7 +119,7 @@ impl Provider for OpenRouterProvider {
                 let code = error_json["error"]["code"]
                     .as_i64()
                     .or_else(|| error_json["error"]["type"].as_str().map(|t| t.len() as i64))
-                    .unwrap_or(0) as i32;
+                    .unwrap_or(0) as u16;
                 let message = error_json["error"]["message"]
                     .as_str()
                     .or_else(|| error_json["error"].as_str())
@@ -128,7 +128,7 @@ impl Provider for OpenRouterProvider {
                 return Err(ProviderError::ApiError { code, message });
             }
             return Err(ProviderError::ApiError {
-                code: status.as_u16() as i32,
+                code: status.as_u16(),
                 message: body,
             });
         }
@@ -160,7 +160,7 @@ impl Provider for OpenRouterProvider {
             let status = response.status();
             let body = response.text().await.unwrap_or_default();
             return Err(ProviderError::ApiError {
-                code: status.as_u16() as i32,
+                code: status.as_u16(),
                 message: body,
             });
         }
